pyproject.toml
[project]
name = "semanticweb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.3.0",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "networkx>=3.5",
    "numpy>=2.3.4",
    "openai>=2.8.0",
    "pandas>=2.3.3",
    "pip>=25.3",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
    "transformers>=4.57.1",
    "langgraph>=0.2.0",
    "langchain-openai>=0.1.0",
    "python-dotenv>=1.2.1",
]
